Revert "fix: cast to int64 instead int32"
This reverts commit 1ecb7ec2f8a1fe8a94a45e63a0a2f40335298ae9.

+++ Doppler.EasyNetQ.HosepipeWorker/HosepipeService.cs
@@ -123,9 +123,9 @@ namespace Doppler.EasyNetQ.HosepipeWorker
                     error.BasicProperties.Headers.Add(_hosepipeSettings.RetryCountHeader, retryCount);
                 }
 
-                error.BasicProperties.Headers[_hosepipeSettings.RetryCountHeader] = (long)retryCount + 1;
+                error.BasicProperties.Headers[_hosepipeSettings.RetryCountHeader] = (int)retryCount + 1;
 
-                if ((long)error.BasicProperties.Headers[_hosepipeSettings.RetryCountHeader] > _hosepipeSettings.MaxRetryCount)
+                if ((int)error.BasicProperties.Headers[_hosepipeSettings.RetryCountHeader] > _hosepipeSettings.MaxRetryCount)
                 {
                     await _busStation.GetBus(service).PublishAsync(error, configure => configure.WithQueueName(_hosepipeSettings.UnsolvedErrorQueueName));
